move AirdropDetail to /airdrop
