+++ src/store/plasma/airdrop/mutations.ts
@@ -1,6 +1,5 @@
-import { AirdropDetail } from "../types"
 import { Airdrop } from "./web3-contracts/Airdrop"
-import { AirdropState } from "./types"
+import { AirdropState, AirdropDetail } from "./types"
 
 export function setAirdropContract(state: AirdropState, payload: Airdrop) {
   state.airdropContract = payload
