+++ src/store/plasma/airdrop/types.ts
@@ -1,4 +1,4 @@
-import { HasPlasmaState, AirdropDetail } from "../types"
+import { HasPlasmaState } from "../types"
 import { Airdrop } from "@/store/plasma/airdrop/web3-contracts/Airdrop"
 import { BareActionContext } from "vuex-typex"
 
@@ -9,6 +9,14 @@ export interface HasAirdropState extends HasPlasmaState {
 export interface AirdropState {
   airdropContract: Airdrop | null,
   usersAirdrops: AirdropDetail[]
+}
+export interface AirdropDetail {
+  airdropID: number,
+  receiver: string,
+  tokenAddress: string,
+  airdropAmount: number,
+  timelock: number,
+  isWithdrew: boolean,
 }
 
 // helper/shorthand for plasma module action context
