+++ src/store/plasma/types.ts
@@ -96,15 +96,6 @@ export enum TierID {
   DEFAULT = 0,
 }
 
-export interface AirdropDetail {
-  airdropID: number,
-  receiver: string,
-  tokenAddress: string,
-  airdropAmount: number,
-  timelock: number,
-  isWithdrew: boolean,
-}
-
 // helper/shorthand for plasma module action context
 export declare type PlasmaContext = BareActionContext<
   PlasmaState,
